+++ Tests/SwiftTileJSONTests/EncodingTests.swift
@@ -10,13 +10,13 @@ import Testing
 @testable import SwiftTileJSON
 
 struct EncodingTests {
-    @Test func customFieldsEncoding() {
+    @Test func extendedFieldsEncoding() {
         let tileJSON = TileJSON(tiles: ["http://a.tileserver.org/{z}/{x}/{y}"])
-        let customFields: [String: Any] = [
+        let extendedFields: [String: Any] = [
             "something_custom": "this is my unique field",
             "another_custom": 42
         ]
-        let extendedTileJSON = ExtendedTileJSON(tileJSON: tileJSON, customFields: customFields)
+        let extendedTileJSON = ExtendedTileJSON(tileJSON: tileJSON, extendedFields: extendedFields)
         
         let jsonData = try! JSONEncoder().encode(extendedTileJSON)
         let decoded = try! JSONSerialization.jsonObject(with: jsonData, options: []) as! [String: Any]