+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## `main`
 * Fix `TileJSON` init parameters
 * Rename `CustomFieldsTileJSON` to `ExtendedTileJSON`
+* Rename `ExtendedTileJSON.customFields` to `extendedFields`.
 
 ## 0.1.1
 * Add more documentation